path-tracer-preprocessor/preprocessor-function/requirements.txt
requests
pygltflib
boto3
pyyaml